backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pyyaml==6.0.2
feedparser==6.0.11
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
apscheduler==3.10.4
google-api-python-client==2.154.0
python-dotenv==1.0.1
